project/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
psycopg2
python-dotenv
supabase 
pandas
numpy